pyproject.toml
[tool.poetry]
name = "usinasoft-back"
version = "0.1.0"
description = "Sistema de Gestão de Usinagem - Backend Django"
authors = ["UsinaSoft Team"]
readme = "README.md"
package-mode = false
license = "Proprietary"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2"
djangorestframework = "^3.14"
django-cors-headers = "^4.3"

[tool.poetry.group.dev.dependencies]
black = "^24.0"
flake8 = "^7.0"
pytest = "^8.0"
pytest-django = "^4.7"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py310']
include = '\.pyi?$'

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "usinasoft.settings"
python_files = ["test_*.py", "*_test.py"]
pythonpath = ["."]
